xkb/servermd: fix bit shifting by CWE-190 Integer Overflow or Wraparound
References:
- https://cwe.mitre.org/data/definitions/190.html

Example real CVE vulns with Integer Overflow:
- https://nvd.nist.gov/vuln/detail/CVE-2022-49748
- https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=08245672cdc6505550d1a5020603b0a8d4a6dcc7

+++ include/servermd.h
@@ -90,11 +90,11 @@ extern _X_EXPORT PaddingInfo PixmapWidthPaddingInfo[];
 
 #define PixmapWidthInPadUnits(w, d) \
     (PixmapWidthPaddingInfo[d].notPower2 ? \
-    (((int)(w) * PixmapWidthPaddingInfo[d].bytesPerPixel +  \
+    ((int)((int)(w) * PixmapWidthPaddingInfo[d].bytesPerPixel +  \
 	         PixmapWidthPaddingInfo[d].bytesPerPixel) >> \
-	PixmapWidthPaddingInfo[d].padBytesLog2) : \
-    ((int)((w) + PixmapWidthPaddingInfo[d].padRoundUp) >> \
-	PixmapWidthPaddingInfo[d].padPixelsLog2))
+	(int)PixmapWidthPaddingInfo[d].padBytesLog2) : \
+    ((int)((int)(w) + PixmapWidthPaddingInfo[d].padRoundUp) >> \
+	(int)PixmapWidthPaddingInfo[d].padPixelsLog2))
 
 /*
  *	Return the number of bytes to which a scanline of the given
